-add nordvpn new api to fetch more data -set foundation to nordvpn wireguard connections -changes number from uint16 to uint32 -add generic to some methods

+++ internal/configuration/sources/env/serverselection.go
@@ -50,7 +50,7 @@ func (s *Source) readServerSelection(vpnProvider, vpnType string) (
 
 	if csv := getCleanedEnv("SERVER_NUMBER"); csv != "" {
 		numbersStrings := strings.Split(csv, ",")
-		numbers := make([]uint16, len(numbersStrings))
+		numbers := make([]uint32, len(numbersStrings))
 		for i, numberString := range numbersStrings {
 			const base, bitSize = 10, 16
 			number, err := strconv.ParseInt(numberString, base, bitSize)
@@ -61,7 +61,7 @@ func (s *Source) readServerSelection(vpnProvider, vpnType string) (
 				return ss, fmt.Errorf("%w: %d must be between 0 and 65535",
 					ErrServerNumberNotValid, number)
 			}
-			numbers[i] = uint16(number)
+			numbers[i] = uint32(number)
 		}
 		ss.Numbers = numbers
 	}
